requirements.txt
Pillow==9.3.0
protobuf==4.21.9
streamlit==1.11.0
google
google.cloud
